top(1): move command mapping to commands.c
This eliminates the difficult to follow mapping of a string list. It
moves numbers from "#define" into (more) debuggable enums. More
generally, it follows the trend of moving more data into a more central
mechanism.

The help output is a little worse: " " is not rendered well, and there
are duplicate entries, but that will be fixed in a followup.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@334988 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f