Add a counter to limit the number of disabled DSs for a mirrored pNFS MDS.
This patch adds a counter that limits the number of disabled mirrored DSs
to mirror level - 1.  It also makes a small change that keeps a Write that
has failed with EACCES when attempted by a client to a DS from disabling
the DS.
This patch only affects the pNFS server.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@335525 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f