allwinner: clkng: Correct mux width and flags
The test for checking if the clock have a mux was inverted and the mask
to calculate the parent index was wrong was wrong too.
It means that upon creation the incorrect parent was resolved as the current
one and upon reparent the switch was never made.

Pointy hat (lots of them): manu


git-svn-id: svn+ssh://svn.freebsd.org/base/head@335581 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f